+++ oap-server/server-receiver-plugin/envoy-metrics-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/envoy/AccessLogServiceGRPCHandler.java
@@ -80,7 +80,7 @@ public class AccessLogServiceGRPCHandler extends AccessLogServiceGrpc.AccessLogS
 
         MetricsCreator metricCreator = manager.find(TelemetryModule.NAME).provider().getService(MetricsCreator.class);
         counter = metricCreator.createCounter(
-            "envoy_als_in_count", "The count of envoy ALS metric received", MetricsTag.EMPTY_KEY,
+            "envoy_als_in_count", "The count of envoy ALS message received", MetricsTag.EMPTY_KEY,
             MetricsTag.EMPTY_VALUE
         );
         histogram = metricCreator.createHistogramMetric(
@@ -103,8 +103,6 @@ public class AccessLogServiceGRPCHandler extends AccessLogServiceGrpc.AccessLogS
 
             @Override
             public void onNext(StreamAccessLogsMessage message) {
-                counter.inc();
-
                 HistogramMetrics.Timer timer = histogram.createTimer();
                 try {
                     if (isFirst) {
@@ -129,6 +127,7 @@ public class AccessLogServiceGRPCHandler extends AccessLogServiceGrpc.AccessLogS
                     switch (logCase) {
                         case HTTP_LOGS:
                             StreamAccessLogsMessage.HTTPAccessLogEntries logs = message.getHttpLogs();
+                            counter.inc(logs.getLogEntryCount());
 
                             for (final HTTPAccessLogEntry log : logs.getLogEntryList()) {
                                 AccessLogAnalyzer.Result result = AccessLogAnalyzer.Result.builder().build();
@@ -143,6 +142,7 @@ public class AccessLogServiceGRPCHandler extends AccessLogServiceGrpc.AccessLogS
                             break;
                         case TCP_LOGS:
                             StreamAccessLogsMessage.TCPAccessLogEntries tcpLogs = message.getTcpLogs();
+                            counter.inc(tcpLogs.getLogEntryCount());
 
                             for (final TCPAccessLogEntry tcpLog : tcpLogs.getLogEntryList()) {
                                 AccessLogAnalyzer.Result result = AccessLogAnalyzer.Result.builder().build();